Split plugins file
